platform/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
python-multipart>=0.0.9
aiofiles>=24.1.0
appium-python-client>=4.0.0
httpx>=0.27.0
